nit: Remove redundant codes - Remove redundant logs - Remove systemcontracts package - Move generated_contracts to consortium package
