+++ core/state_processor.go
@@ -18,7 +18,6 @@ package core
 
 import (
 	"fmt"
-	"github.com/ethereum/go-ethereum/core/systemcontracts"
 	"math/big"
 
 	"github.com/ethereum/go-ethereum/common"
@@ -72,8 +71,6 @@ func (p *StateProcessor) Process(block *types.Block, statedb *state.StateDB, cfg
 	if p.config.DAOForkSupport && p.config.DAOForkBlock != nil && p.config.DAOForkBlock.Cmp(block.Number()) == 0 {
 		misc.ApplyDAOHardFork(statedb)
 	}
-	// update system contracts before processing EVM
-	systemcontracts.UpgradeBuildInSystemContract(p.config, block.Number(), statedb)
 
 	blockContext := NewEVMBlockContext(header, p.bc, nil, publishEvents...)
 	vmenv := vm.NewEVM(blockContext, vm.TxContext{}, statedb, p.config, cfg)
