--- core/systemcontracts/const.go
@@ -1,9 +1,0 @@
-package systemcontracts
-
-const (
-	ValidatorContract = "0x0000000000000000000000000000000000001000"
-	SlashContract     = "0x0000000000000000000000000000000000001001"
-
-	ConsortiumV2ValidatorContractCode = ""
-	ConsortiumV2SlashContractCode     = ""
-)
